Added a comments section to the project page. You need to be signed in to add comments.
I added step_defs for users

I created a comments model and controller for posting comments. The controler
only has a create action. It may need a destroy one in the future.

I added a two utility methods to the User
  full_name - simply concats the first and last names
  location - This returns the string "New York, NY" for now. The user will need
    a location in order to meet the design mockup. Not sure that is what will be
    wanted?

May need to add a story for moderating comments. Perhaps comments need to be
approved. Or all are visible but can be deleted by and admin etc.